Improve Closure minification
