+++ src/closure-annotations.js
@@ -1,0 +1,210 @@
+// WebIDL
+
+/**
+ * @suppress {duplicate}
+ */
+var WrapperObject;
+/**
+ * @suppress {duplicate}
+ */
+var getCache;
+/**
+ * @suppress {duplicate}
+ */
+var wrapPointer;
+/**
+ * @suppress {duplicate}
+ */
+var castObject;
+/**
+ * @suppress {duplicate}
+ */
+var destroy;
+/**
+ * @suppress {duplicate}
+ */
+var compare;
+/**
+ * @suppress {duplicate}
+ */
+var getPointer;
+/**
+ * @suppress {duplicate}
+ */
+var getClass;
+/**
+ * @suppress {duplicate}
+ */
+var ensureCache;
+/**
+ * @suppress {duplicate}
+ */
+var ensureString;
+/**
+ * @suppress {duplicate}
+ */
+var ensureInt8;
+/**
+ * @suppress {duplicate}
+ */
+var ensureInt16;
+/**
+ * @suppress {duplicate}
+ */
+var ensureInt32;
+/**
+ * @suppress {duplicate}
+ */
+var ensureFloat32;
+/**
+ * @suppress {duplicate}
+ */
+var ensureFloat64;
+/**
+ * @suppress {duplicate}
+ */
+var VoidPtr;
+
+// Various Emscripten-specific global variables
+var tempI64;
+var tempDouble;
+
+/**
+ * GL functions (https://github.com/emscripten-core/emscripten/commit/97a464a654fdadf5dfb8aa082b48516e6bf8d402#commitcomment-25520648)
+ */
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glDrawArrays;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glDrawElements;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glActiveTexture;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glEnable;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glDisable;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glTexEnvf;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glTexEnvi;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glTexEnvfv;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glGetIntegerv;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glIsEnabled;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glGetBooleanv;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glGetString;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glCreateShader;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glShaderSource;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glCompileShader;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glAttachShader;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glDetachShader;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glUseProgram;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glDeleteProgram;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glBindAttribLocation;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glLinkProgram;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glBindBuffer;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glGetFloatv;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glHint;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glEnableVertexAttribArray;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glDisableVertexAttribArray;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _emscripten_glVertexAttribPointer;
+
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _glTexEnvf;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _glTexEnvi;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _glTexEnvfv;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _glGetTexEnviv;
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var _glGetTexEnvfv;
+
+var _glutPostRedisplay = function() {};
+
+/**
+ * @suppress {undefinedVars}
+ */
+var emscripten_source_map;
