+++ src/closure-defined-fs-annotation.js
@@ -1,0 +1,39 @@
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var Browser;
+
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var SDL;
+
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var GL;
+
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var GLctx;
+
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var SOCKFS;
+
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var Browser;
+
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var __read_sockaddr;
+
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var DNS;
