+++ src/closure-undefined-fs-annotation.js
@@ -1,0 +1,4 @@
+/**
+ * @suppress {duplicate, undefinedVars}
+ */
+var FS;
