+++ src/shell.js
@@ -227,14 +227,12 @@ if (ENVIRONMENT_IS_WEB || ENVIRONMENT_IS_WORKER) {
   } else if (document.currentScript) { // web
     scriptDirectory = document.currentScript.src;
   }
-#if MODULARIZE
-#if MODULARIZE_INSTANCE == 0
+#if MODULARIZE && MODULARIZE_INSTANCE == 0
   // When MODULARIZE (and not _INSTANCE), this JS may be executed later, after document.currentScript
   // is gone, so we saved it, and we use it here instead of any other info.
   if (_scriptDir) {
     scriptDirectory = _scriptDir;
   }
-#endif
 #endif
   // blob urls look like blob:http://site.com/etc/etc and we cannot infer anything from them.
   // otherwise, slice off the final part of the url to find the script directory.
