+++ tools/shared.py
@@ -1239,6 +1239,10 @@ class SettingsManager(object):
     def __getitem__(self, key):
       return self.attrs[key]
 
+    @classmethod
+    def target_environment_may_be(self, environment):
+      return self.attrs['ENVIRONMENT'] == '' or environment in self.attrs['ENVIRONMENT'].split(',')
+
   __instance = None
 
   @staticmethod
@@ -2434,11 +2438,28 @@ class Building(object):
         logger.error('Cannot run closure compiler')
         raise Exception('closure compiler check failed')
 
+      # Closure annotations file contains suppressions and annotations to different symbols
+      CLOSURE_ANNOTATIONS = ['--js', path_from_root('src', 'closure-annotations.js')]
+
+      if not Settings.ASMFS:
+        # If we have filesystem disabled, tell Closure not to bark when there are syscalls emitted that still reference the nonexisting FS object.
+        if not Settings.FILESYSTEM:
+          CLOSURE_ANNOTATIONS += ['--js', path_from_root('src', 'closure-undefined-fs-annotation.js')]
+
+        # If we do have filesystem enabled, tell Closure not to bark when FS references different libraries that might not exist.
+        if Settings.FILESYSTEM and not Settings.ASMFS:
+          CLOSURE_ANNOTATIONS += ['--js', path_from_root('src', 'closure-defined-fs-annotation.js')]
+
+      # Closure externs file contains known symbols to be extern to the minification, Closure
+      # should not minify these symbol names.
       CLOSURE_EXTERNS = path_from_root('src', 'closure-externs.js')
       NODE_EXTERNS_BASE = path_from_root('third_party', 'closure-compiler', 'node-externs')
       NODE_EXTERNS = os.listdir(NODE_EXTERNS_BASE)
       NODE_EXTERNS = [os.path.join(NODE_EXTERNS_BASE, name) for name in NODE_EXTERNS
                       if name.endswith('.js')]
+      NODE_EXTERNS = [path_from_root('src', 'node-externs.js')] + NODE_EXTERNS
+      V8_EXTERNS = [path_from_root('src', 'v8-externs.js')]
+      SPIDERMONKEY_EXTERNS = [path_from_root('src', 'spidermonkey-externs.js')]
       BROWSER_EXTERNS_BASE = path_from_root('third_party', 'closure-compiler', 'browser-externs')
       BROWSER_EXTERNS = os.listdir(BROWSER_EXTERNS_BASE)
       BROWSER_EXTERNS = [os.path.join(BROWSER_EXTERNS_BASE, name) for name in BROWSER_EXTERNS
@@ -2451,27 +2472,34 @@ class Building(object):
               '-Xmx' + (os.environ.get('JAVA_HEAP_SIZE') or '1024m'), # if you need a larger Java heap, use this environment variable
               '-jar', CLOSURE_COMPILER,
               '--compilation_level', 'ADVANCED_OPTIMIZATIONS',
-              '--language_in', 'ECMASCRIPT5',
-              '--externs', CLOSURE_EXTERNS,
-              # '--variable_map_output_file', filename + '.vars',
-              '--js', filename, '--js_output_file', outfile]
-      for extern in NODE_EXTERNS:
-        args.append('--externs')
-        args.append(extern)
-      for extern in BROWSER_EXTERNS:
-        args.append('--externs')
-        args.append(extern)
+              '--language_in', 'ECMASCRIPT5']
+      args += CLOSURE_ANNOTATIONS
+      args += ['--externs', CLOSURE_EXTERNS,
+               '--js_output_file', outfile]
+
+      if Settings.target_environment_may_be('node'):
+        for extern in NODE_EXTERNS:
+          args.append('--externs')
+          args.append(extern)
+      if Settings.target_environment_may_be('shell'):
+        for extern in V8_EXTERNS + SPIDERMONKEY_EXTERNS:
+          args.append('--externs')
+          args.append(extern)
+      if Settings.target_environment_may_be('web') or Settings.target_environment_may_be('worker'):
+        for extern in BROWSER_EXTERNS:
+          args.append('--externs')
+          args.append(extern)
       # Closure compiler needs to know about all exports that come from the asm.js/wasm module, because to optimize for small code size,
       # the exported symbols are added to global scope via a foreach loop in a way that evades Closure's static analysis. With an explicit
       # externs file for the exports, Closure is able to reason about the exports.
-      if Settings.MODULE_EXPORTS:
+      if Settings.MODULE_EXPORTS and not Settings.DECLARE_ASM_MODULE_EXPORTS:
         # Generate an exports file that records all the exported symbols from asm.js/wasm module.
         module_exports_suppressions = '\n'.join(['/**\n * @suppress {duplicate, undefinedVars}\n */\nvar %s;\n' % i for i in Settings.MODULE_EXPORTS])
         exports_file = configuration.get_temp_files().get('_module_exports.js')
         exports_file.write(module_exports_suppressions.encode())
         exports_file.close()
 
-        args.append('--externs')
+        args.append('--js')
         args.append(exports_file.name)
       if Settings.IGNORE_CLOSURE_COMPILER_ERRORS:
         args.append('--jscomp_off=*')
@@ -2479,6 +2507,7 @@ class Building(object):
         args += ['--formatting', 'PRETTY_PRINT']
       if os.environ.get('EMCC_CLOSURE_ARGS'):
         args += shlex.split(os.environ.get('EMCC_CLOSURE_ARGS'))
+      args += ['--js', filename]
       logger.debug('closure compiler: ' + ' '.join(args))
       proc = run_process(args, stderr=PIPE, check=False)
       if proc.returncode != 0:
